+++ web/src/test/auto/dom/test_utils.ts
@@ -82,15 +82,3 @@ export class DynamicElements {
     return editable.id;
   }
 }
-
-export const coreurl = '/build/engine/obj/core-adapter/import/core';
-
-export async function loadKeyboardBlob(uri: string) {
-  const response = await fetch(uri);
-  if (!response.ok) {
-    throw new Error(`HTTP ${response.status} ${response.statusText}`);
-  }
-
-  const buffer = await response.arrayBuffer();
-  return new Uint8Array(buffer);
-}
